refactor: organize code into schemas/, utils/, and types/ folders
